requirements.txt
aws-msk-iam-sasl-signer-python==1.0.1
boto3==1.35.90
botocore==1.35.90
click==8.1.8
colorama==0.4.6
jmespath==1.0.1
kafka-python==2.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.4
six==1.17.0
urllib3==1.26.20
websocket-client==1.8.0
